requirements.txt
Babel==2.8.0
Brotli==1.0.9
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2
django-address==0.2.3
django-admin-interface==0.14.2
django-appconf==1.0.4
django-colorfield==0.3.2
django-extensions==3.0.9
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-imagekit==4.0.2
django-jazzmin==2.4.1
django-phone-field==1.8.1
django-phonenumber-field==4.0.0
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
django-shell-plus==1.1.7
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
django-rest-passwordreset
gunicorn==20.0.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
pilkit==2.0
Pillow==5.2.0
psycopg2==2.7.4
psycopg2-binary==2.8.6
PyJWT==1.6.4
python-decouple==3.3
pytz==2018.5
requests==2.24.0
simplejson==3.17.2
six==1.11.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.2.0
